refactor: densify five more 380-line app widget hosts for wave 21
Move remaining cargo-cap, tech chrome, new-game duplicate-slot, map-options,
and standing-chip cases into sibling suites so Slice D hosts stay under 360.
